Have custom coercer-class
